+++ figlet4s-core/src/main/scala/com/colofabrix/scala/figlet4s/rendering/HorizontalMergeRules.scala
@@ -1,14 +1,10 @@
 package com.colofabrix.scala.figlet4s.rendering
 
 import cats.implicits._
-import com.colofabrix.scala.figlet4s.figfont._
 import com.colofabrix.scala.figlet4s.figfont.FIGfontParameters._
 import com.colofabrix.scala.figlet4s.rendering.MergeAction._
 import com.colofabrix.scala.figlet4s.rendering.Rendering._
-import com.colofabrix.scala.figlet4s.options.{
-  HorizontalLayout => ClientHorizontalLayout, PrintDirection => ClientPrintDirection,
-  Justification => ClientJustification, RenderOptions,
-}
+import com.colofabrix.scala.figlet4s.options.{ HorizontalLayout => ClientHorizontalLayout, RenderOptions }
 
 /**
  * Merging rules for horizontal appending of characters
@@ -25,45 +21,7 @@ private[figlet4s] object HorizontalMergeRules {
     layout2mergeStrategy(options.font.header.hardblank)(chosenLayout)
   }
 
-  def applyPrintDirection(options: RenderOptions, text: String): String =
-    options.printDirection match {
-      case ClientPrintDirection.LeftToRight => text
-      case ClientPrintDirection.RightToLeft => text.reverse
-      case ClientPrintDirection.FontDefault =>
-        options.font.settings.printDirection match {
-          case FIGfontParameters.PrintDirection.LeftToRight => text
-          case FIGfontParameters.PrintDirection.RightToLeft => text.reverse
-        }
-    }
-
-  def applyFlushing(options: RenderOptions, outputLines: Seq[SubColumns]): Seq[SubColumns] =
-    outputLines.map { columns =>
-      val lines = columns.toSublines
-      val flushed =
-        options.justification match {
-          case ClientJustification.FlushLeft   => lines
-          case ClientJustification.Center      => alignCenter(lines, options.maxWidth)
-          case ClientJustification.FlushRight  => alignRight(lines, options.maxWidth)
-          case ClientJustification.FontDefault => lines
-        }
-      flushed.toSubcolumns
-    }
-
   //  Support  //
-
-  /** Align SubLines in the center of the given width */
-  private def alignCenter(input: SubLines, maxWidth: Int): SubLines = {
-    val halfSpaces   = (maxWidth - input.width) / 2.0
-    val leftPadding  = " " * Math.ceil(halfSpaces).toInt
-    val rightPadding = " " * Math.floor(halfSpaces).toInt
-    input.map(leftPadding + _ + rightPadding)
-  }
-
-  /** Align SubLines at the right of the given width */
-  private def alignRight(input: SubLines, maxWidth: Int): SubLines = {
-    val leftPadding = " " * (maxWidth - input.width - 1)
-    input.map(leftPadding + _)
-  }
 
   /** Function that smushes two characters */
   private type SmushingStrategy = (Char, Char) => Option[Char]